Fixing signed/unsigned comparison warnings in ZeroD

+++ cantera18/branches/scons/Cantera/src/zeroD/Reactor.h
@@ -103,7 +103,7 @@ namespace CanteraZeroD {
         virtual void updateState(doublereal* y);
 
         virtual size_t nSensParams();
-        virtual void addSensitivityReaction(int rxn);
+        virtual void addSensitivityReaction(size_t rxn);
 
         virtual std::string sensParamID(int p) { return m_pname[p]; }
